container_utils: wait for device_jail_fs export in init
device_jail_fs may take some time to mount if the fuse module
is not loaded yet, so there is a potential race when starting up
a container for the first time (and also if ARC++ has not been
started). Here we use a post-start script to ensure that the
process is not considered "started" until its mount shows up in
the /proc/mounts table.

BUG=chromium:728308
TEST=insert delay into device_jail_fs initialization sequence
  after it starts but before it registers the fuse filesystem,
  ensure that the device filesystem is still mounted inside the
  container when it starts up, verify that it works if ARC++ was
  not started first (and thus the fuse module hasn't been loaded
  prior)

Change-Id: I4cf272506d3be3f09e8ddbe884769f24539cf48a
Reviewed-on: https://chromium-review.googlesource.com/521802
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>