midis: Perform minor test code refactoring.
A lot of the test code shares common sources. This patch moves
the common sources into a common library. We also move the tests into a
separate tests/ directory for better organization.

We also use the patch to move all the tests to use the common testrunner
used by the rest of ChromeOS as well as to put all the tests in a single
executable.

BUG=chromium:701005
TEST=cros_run_unit_tests --board chell --packages "midis"
CQ-DEPEND=CL:528441

Change-Id: Iade9bf0a13221fe83f1b1375df74e4952e0e8a90
Reviewed-on: https://chromium-review.googlesource.com/527431
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>