metrics: Add bulk update API.
Use metrics_uploader to define
USE_MERICS_UPLOADER and in turn conditionally compile the
new API.

Note once the support for chrome is in, we can retire
this conditional flag.

The API Is SendRepeatedToUMA(), the underneath mechanism is
the same as SendToUMA().

This extension allows sending multiple identical histogram samples
in a single call.

It is introduced to reduce overhead in the jetstream QoS metrics
use case.

Effectively, SendToUMA(...) is the same as SendRepeatedToUMA(..., 1).

In order to support this API, class metric_sample is augmented with
the private member "int num_samples_", in order to have
backward compatibility, the serialization conditionally append the log
with the num_samples_ if and only if num_samples_ is greater than 1.
The deserialization of histogram in turn checks the number of field,
and conditionally updates the value of num_samples_.

All the other API are kept the same as before w.r.t to
serialization/deserialization.

BUG=b:62483253
TEST=Add unit test for the new api and the test passed.
Change-Id: Ib0f577ba76f30f88b46fbdd5077cf2b6491918ae
Reviewed-on: https://chromium-review.googlesource.com/528368
Commit-Ready: Mu Lin <mxt@google.com>
Tested-by: Mu Lin <mxt@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mu Lin <mxt@google.com>