midis: Add client library functions
This patch adds a few helper functions that clients can use
to make establishing and maintaining a connection with the MIDI server
easier. It implements a few primitives like reading messages and
headers, so that the client doesn't need to reimplement these.

Having a reference implementation means that minor changes in the
protocol and message formats can be hidden from the client.

We also update the client tests to use the new library routines.

BUG=chromium:701005
TEST=cros_run_unit_tests --board chell --packages "midis"
CQ-DEPEND=CL:483431

Change-Id: Iafb83bc043601fd669229e5438ae657ca59ccad1
Reviewed-on: https://chromium-review.googlesource.com/528522
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>