cryptohome: Record free space and xattr size
When the ecryptfs->ext4crypto fails with an ENOSPC record UMA stats
including the initial free space when the migration started as well as
the free space at time of failure.  If the error occurs during an
attempt to set an xattr, also record the total xattr size of the
existing attributes on the file in addition to the xattr that was not
successfully set.

TEST=cros_workon_make --board=samus --test cryptohome
BUG=chromium:731505

Change-Id: Iac282ca83109dff7b81680aa04f01df8e62dc4a5
Reviewed-on: https://chromium-review.googlesource.com/527742
Commit-Ready: Dan Spaid <dspaid@chromium.org>
Tested-by: Dan Spaid <dspaid@chromium.org>
Reviewed-by: Dan Spaid <dspaid@chromium.org>