init: Install cros_payloads in a clean and safer way.
We found that in early, for example proto, builds, device may be in
unstable state and reboot unexpectedly during first boot. This implies
cros_payload installation may go wrong if the machine reboots before the
changes to system disk is flushed.

As a result, we want to revise the payload installation to its own
service, cros-payloads.conf, so it won't rely on factory stuff.
Also, change the "removal of install stubs" to be postponed to few
seconds after sync(1) is invoked. This should allow re-installation if
system is unstable in early stage.

BUG=chromium:731520
TEST=Manually installed on system and verified cros_payload resources
     were installed.

Change-Id: Ia7b702997de7168fea56eed98525b74e81791bef
Reviewed-on: https://chromium-review.googlesource.com/527548
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>