login: Remove callback for Start/StopArc in SessionManagerImpl.
The callback closures are used only for testing.
It used be valuable, because it was only the way to see
if the D-Bus method is completed successfully or not.

Now, we switched tool-generated interface, so it explictly
returns boolean indicating if the method is successfully
done or not, which means, we no longer need the closure.

Note that, EXPECT_TRUE(arc_setup_completed_) is identical
to EXPECT_TRUE(impl_.StartArcInstance()).
EXPECT_FALSE(arc_setup_completed_) is identical to
EXPECT_FALSE(android_container_.running()) in the testing code.

BUG=None
TEST=Ran cros_workon_make --test

Change-Id: Iae62567e2c6321cfacf4f80cc4e8c3392e56b52b
Reviewed-on: https://chromium-review.googlesource.com/535261
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>