login: Use ON_CALL for un-interested mock expectations.
For testing perspective, MockSystemUtil is an object whose
behavior needs to be set to avoid real side-effect,
than a target object whose method calls need to be verified.

Similary, UserProfileServiceFactory is an object to inject
mock instances to be verified, rather than an object to be verified.

In such a case, ON_CALL is better to be used because it's behavior wouldn't
be cleared by VerifyAndClearExpectations().
So, once they are set in beginning (i.e. SetUp()), they can be left.

Along with the change, these are switched to NiceMock to avoid
verbose "Unexpected call" warnings, because these are expected.

BUG=None
TEST=Ran cros_workon_make --test chromeos-login

Change-Id: I8452186827bd76e6c91fc096b332ca481ac414e9
Reviewed-on: https://chromium-review.googlesource.com/538539
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>