chaps: don't manually unmarshal protos in adaptor
The libbrillo bindings do have support for taking protos directly,
if they are subclasses of MessageLite (which all chaps protots are).
This means we don't have to take a vector and then unmarshal it;
the libbrillo bindings will do that if the method we register takes
a MessageLite subclass.

BUG=chromium:719073
TEST=unit tests, list slots, tokens and mechanisms with pkcs11-tool

Change-Id: I02e4b999f657a8a59aae9e43b65ae50d17e72b01
Reviewed-on: https://chromium-review.googlesource.com/538855
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>