arc-networkd: Import DNS response parser from Chromium
This isn't in libchrome, and the original code pulls in a long string
of dependencies.  Import a trimmed-down version of the parser from
Chromium rev 9d2ed1642ac8b258.

The functions that are part of this fork use the same API as the
original Chromium functions, so if the net/dns code is ever integrated
into libchrome, we will be able to delete the arc-networkd/dns/ tree.

BUG=b:36449008
TEST=compile-tested only

Change-Id: Ic7625d73bd44ce6ceec2e920bb11f5e118662ab3
Reviewed-on: https://chromium-review.googlesource.com/539180
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@google.com>