power: suspend_stress_test: Don't check s0ix if driver suspend fails
In the case where the system fails to suspend due to a driver failure,
the system will not even try to enter s0ix, and then s0ix counter will
not increment, as expected. In such situation, however, we should not
categorise this as s0ix failure, because the system actually never tried
to enter s0ix. So we should not display "s0ix counters did not
increment..." and should not increment s0ix_errors.

Thus change the code so that s0ix counters are checked only if the
system actually did try to enter s0ix.

BUG=b:36630881
TEST=Try S0ix test with driver failures on Soraka

Change-Id: Idb216157bf6073c883b2a30715f78993bd80c9e6
Reviewed-on: https://chromium-review.googlesource.com/544016
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>