shill: reduce number of bucket for *.ExpiredLeaseLengthSeconds metrics
shill records *.ExpiredLeaseLengthSeconds metrics for various network
types (e.g. Ethernet, WiFi). Each of these metrics is recorded as a
count histogram and tracks the DHCP lease length for a network of a
particular type when the lease expires without the DHCP client being
able to renew it.

The maximum of the count histogram for *.ExpiredLeaseLengthSeconds is
set to 604800 (the total number of seconds in 7 days) while the number
of bucket is set to 604801, which far exceeds the maximum bucket count
(16384) specified by the Chrome metrics library.

[1] states: "By default count histograms bucket sizes scale
exponentially so you can get fine granularity when the numbers are small
yet still reasonable resolution for larger numbers. The macros default
to 50 buckets (or 100 buckets for histograms with wide ranges) which is
appropriate for most purposes."

This CL reduces the bucket count for *.ExpiredLeaseLengthSeconds to
100 as recommended by [1].

[1] https://chromium.googlesource.com/chromium/src.git/+/HEAD/tools/metrics/histograms/README.md

BUG=chromium:738122
TEST=Run unit tests.

Change-Id: I0bf49e5d1e316efebe6de6ea5b078588c192b57b
Reviewed-on: https://chromium-review.googlesource.com/557297
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>