metrics: Fix permissions error on vmlog dir.
The default permissions on this log directory were too restrictive
(0700, owned by root), and chrome://system and feedback reports are
unable to read this data. This relaxes the permissions so that other
users can read this information (there's nothing sensitive in here, so
it should be fine).

BUG=chromium:739490
TEST=Deploy, Checked directory perms and chrome://system

Change-Id: Ib0eaa40dba8aa87cd05657b14651595f5ef0e3c0
Reviewed-on: https://chromium-review.googlesource.com/560073
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>