debugd: Don't set half-constructed command lines.
If |sandboxing_| is true but |user_| or |group_| are not set, the
cmdline will be incomplete, and subsequent calls to Init() will make
things worse. Same thing if |set_capabilities_| is true but
|sandboxing_| is false or |user_| is root.

Fix this by returning false before the cmdline is built.

BUG=chromium:739736
TEST=Chrome OS boots, chrome:system works.

Change-Id: I3eb01428b796097b434612c049240d38d97296fc
Reviewed-on: https://chromium-review.googlesource.com/561638
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>