installer: fix when IS_RECOVERY_INSTALL is not set
The installer was evaluating the IS_RECOVERY_INSTALL environment variable
in a -u environment, making it fail if the variable is undeclared (which
is the case if the installer is run normally):

./chromeos-install: 1: IS_RECOVERY_INSTALL: parameter not set

Fix this by making the variable expand to the empty string if not
declared.

Fixes c4972eb847 "Report recovery metrics to UMA after recovery".

BUG=chromium:740042
TEST=chromeos-install completes successfully when run as-is.

Change-Id: Id0fe2d653cff4e1286670be57f6df437423504f2
Reviewed-on: https://chromium-review.googlesource.com/563277
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>