authpolicy: Remove misleading logs from smbclient
Authpolicyd logs contained lines like
  smbclient failed with exit code 1
that did not actually indicate an error. What happens is that the
last file that's being downloaded doesn't exist. This is fine,
though.

This CL appends an 'exit;' to the smbclient command, basically a
no-op, but it makes sure the exit code is 0. This gets rid of the
erroneous logs and simplifies logic.

BUG=chromium:740602
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy

Change-Id: I44099a297d5b5c64dbff940074f0e8b7525caf33
Reviewed-on: https://chromium-review.googlesource.com/567148
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>