login: Avoid persisting policy data when no change detected.
When device is shut down, SessionManagerImpl::Finalize is called
as the session is destroyed. The method calls PolicyStore::Persist
to save the policy data.

In this CL the policy data is cached and is persisted to disk
only in case the data is changed. This doesn't avoid persisting
when policy fetch is done because the timestamp is included in
the data, but allows to avoid the writing to disk at shut down.

BUG=chromium:741019
TEST=Manual

Change-Id: Idde3cb666fe6e18b778dad4b5d732c88d32d4fac
Reviewed-on: https://chromium-review.googlesource.com/567510
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>