platform2: Set 1 raster thread when on big_little.
On Kevin it was determined it was better to use only one raster:
chrome-os-partner:56586

The same reason why that'd be better should apply to all the big.LITTLE
boards.

This CL sets the number of raster threads to 1 to all the big.LITTLE
boards.

BUG=chrome-os-partner:56586
TEST=emerge-eve libchromeos-ui

Change-Id: Icf69a51717a72be1df62a529c5e812115a637a91
Reviewed-on: https://chromium-review.googlesource.com/571321
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>