shill: Add SetFixedIpParams() device setting
Allow callers to specify that an interface's IP parameters should not be
changed.  This will be used for the ARC network bridge interface, which
is not configured through shill.  When ARC becomes the default service,
we want shill to change the systemwide DNS and routing configuration but
we do not want it to manipulate the interface's IP configuration
(address, prefix, broadcast, or interface up/down state).

BUG=chromium:696865
TEST=unit tests

Change-Id: I5908c2a61ba2d66c79948610e046f7aa62f84fb7
Reviewed-on: https://chromium-review.googlesource.com/573642
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>