shill: Allow --arc-device to be specified on command line
The ARC device is a network interface that connects the host system
with an ARC container.  shill needs to know which interface it is so
that it can set up ARC based VPNs.  The command line option allows the
interface to be changed from the upstart scripts, which is where much
of the other ARC network / bridge configuration takes place.

The IP configuration of the ARC device will not be managed by shill.
Absent an ARC VPN, it will behave similarly to a blacklisted device.

By default this will use the interface name "arcbr0" instead of "br0",
and the cheets-scripts will be updated accordingly.  This way, a user
who is not running ARC will not see surprising behavioral differences
between a bridge named "br0" and a bridge named "br1".

BUG=chromium:696865
TEST=use --arc-device=arcbr0 instead of adding it to the blacklist, and
     verify that shill leaves its IP configuration alone

Change-Id: I2b648f93520ba7dc00ef171641e01607917e3a6c
Reviewed-on: https://chromium-review.googlesource.com/573643
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>