power: Don't suspend console when using suspend_to_idle
Suspend_on_idle on Intel systems is intended to take the system
into a low power state S0ix instead of S3.

Intel identified that not suspending the UART can cause failure to
go into S0ix state (via suspend-to-idle) and thus rendering system
stay in higher package power state such as PC10, because a non
suspended UART can cause events that affect PMC's decision to enter
S0ix. Hence this patch suspends the console when using Suspend-to-
idle.

The suspend and resume kernel logs are still going to be
available in dmesg / syslog. The logs will also eventually
appear on the UART itself  (but will be printed after a delay
because the messages are saved in a buffer and printed after
the UART is resumed).

BUG=b:63737106
TEST=check that console is suspended when using S0ix, and not suspended
when using S3.

Change-Id: I5b0b28e5e19fe181fa6e58ad67b2f443934b8832
Reviewed-on: https://chromium-review.googlesource.com/580674
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Rajneesh Bhardwaj <rajneesh.bhardwaj@intel.com>