shill: Fix bogus warnings in UpdateFromIPConfig on IPv6-only networks
If a network is IPv6-only, or if the IPv6 configuration occurs
before the DHCPv4 configuration, Connection::UpdateFromIPConfig()
will be run against |Device::ip6config_| instead of
|Device::ipconfig_|.  This does not happen on normal mixed
IPv4+IPv6 networks so it is somewhat of a corner case.
UpdateFromIPConfig() makes two assumptions that are only valid on
IPv4:

1) That a broadcast address exists.  IPv6 only uses multicast.

2) That the default gateway will have the same prefix as our global
scope address.  On IPv6 the gateway address is typically a link-local
address.

Disable these two warnings for IPv6.

BUG=chromium:727242
TEST=manually hack device.cc to delay IPv4 address assignment, then
     check the logs to see what happens when UpdateFromIPConfig() is
     run on an IPv6 configuration
TEST=unit tests

Change-Id: Id343e49f363ab5efaa5be0987a3f138d6279a4ec
Reviewed-on: https://chromium-review.googlesource.com/580642
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>