shill: factor out netlink socket opening function
This code is copied into two initialization paths, with a TODO
to refactor it into a single function. Because we want to use
SOCK_DIAG netlink sockets, it would be better to move this code
now, and then we can use it for future netlink socket types.

BUG=chromium:221940,chromium:726815
TEST=unit tests

Change-Id: Ifd03d4258fd65d385186ae4db42850fbdbf4e473
Reviewed-on: https://chromium-review.googlesource.com/582333
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>