shill: test-scripts: stop looking for 'Favorites'
These died with flimflam a long time ago, in a galaxy far away.

For printing services, it's probably useful to use the asterisk for
notating IsActive. Do that.

For removing services, I don't know what we want to do. But we surely
don't want the "Favorite" property, so kill it. (Note that "remove" is
still broken.)

BUG=chromium:735524
TEST=`./test-flimflam services` gives a reasonable dump

Change-Id: Ieb82028635d82572364dddeb4c56d61823c10cdc
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/583891
Reviewed-by: Ben Chan <benchan@chromium.org>