power: Track keyboard backlight brightness as percent.
Make KeyboardBacklightController track the current
brightness as a percentage instead of as a hardware level
and not cache the maximum level. This isn't intended to have
any effect right now, but it'll make it easy to support
changes to the underlying backlight device in the future.

BUG=chromium:749539
TEST=existing tests pass

Change-Id: I784f4e3dd23c353f81b647e98c92bddaa93f800e
Reviewed-on: https://chromium-review.googlesource.com/591953
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>