midis: Mojo bootstrap code
Establish a Mojo IPC channel with ARC++. Use the first interface,
MidisManagerGetter.Connect(), to initialize and return a handle to
a second interface (MidisManager) which will be initialized per-client,
and will handle all the communication between the client and the server.

MidisManagerGetter.Connect() currently doesn't do anything. The
MidisManager functionality will come in a subsequent CL.

This CL basically establishes the first MidisManagerGetter interface
(initiated in Chrome's ArcBridgeHost after a call from ARC++), and
returns it to the caller.

BUG=chromium:701005
TEST=midis can connect to ARC++

Change-Id: I566413152009838aa610e1ad6c98057d9f79199d
Reviewed-on: https://chromium-review.googlesource.com/594679
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>