cros-disks: Fix GetBootDevicePath in unit tests
rootdev() returns 0 on success, so GetBootDevicePath has been testing
the result backwards.  I think this has been working in the past because
rootdev sometimes sets the path even when it fails and most of the tests
didn't depend on having the correct boot device.  The incorrect logic
started failing when the chroot is on a loopback device for x86_64
boards.

BUG=chromium:748665, chromium:730144
TEST=Sent to trybots with http://crrev.com/c/553462 to confirm inside a
  loopback-based chroot.

Change-Id: Id6ed829667f0055ce6e6ddde292bad18ddc253a3
Reviewed-on: https://chromium-review.googlesource.com/596408
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>