authpolicy: Log seccomp for verbose debug level
Enables logging of seccomp filter failures for the 'kVerbose' debug
level. This level can be set with the crosh command
'authpolicy_debug 3' or through the D-Bus call SetDefaultLogLevel(3).

The AuthPolicy daemon autotest needs a way to enable seccomp logs in
order to print the offending syscall. One way of doing this is to
write certain flags to /etc/authpolicyd_flags, but since DUTs don't
always have rootfs verification disabled, this is not an option.
With this CL, the daemon can use the D-Bus call.

See also CL:600408,CL:601788

BUG=chromium:687177
TEST=test_that -b chell $DUT_IP enterprise_AuthPolicyDaemon

Change-Id: Ia7aac601ec6e5d916bcb847c42305dd8b60959be
Reviewed-on: https://chromium-review.googlesource.com/602128
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>