debugd: fix GetPerfOutput error handling
It was unclear what should actually be returned by GetPerfOutput
in the event of a failure in the perf process itself. This
changes failure to initialize the process to a D-Bus error, and
runtime failures are still returned in the status field, but with
no D-Bus error. This also aligns the return value of the adaptor
with the presence or absence of an error.

BUG=chromium:752431
TEST=unit tests, dbus-send, platform_DebugDaemonGetPerfData

Change-Id: Id51d8755d304b12aad8bb38afc20825c5bc42590
Reviewed-on: https://chromium-review.googlesource.com/602503
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>