login: add flag for background blur.
This CL adds the "--enable-background-blur" flag to the Chrome command
line if the "background_blur" USE flag was set for this board. That
USE flag is found in /etc/ui_use_flags.txt, which is set based on the
USE flags for that board during build time.

BUG=chromium:749145
TEST=Deploy login_manager to Eve and check that command line include flag is used.

Change-Id: Ibe95446c5e7717c14be37afc9ac7fbdfde4506e4
Reviewed-on: https://chromium-review.googlesource.com/603048
Commit-Ready: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>