cryptohome: Migrate ephemeral to ext4 on sparse file on tmpfs
In order to start ARC++ (Android container) in ephemeral mode,
it requires ext4 filesystem.
To adapt this for all-in-RAM ephemeral case, cryptohome creation
changed from just tmpfs mount, to creation of a sparse file there,
create a loop device on it and finally mount it as ext4 cryptohome.

BUG=chromium:774036
TEST=test_that platform_CryptohomeMount
TEST=test_that login_Cryptohome.*
TEST=cros_workon_make --board=${BOARD} chromeos-base/cryptohome --test
TEST=Manually launch public and guest session
CQ-DEPEND=CL:718696

Change-Id: Ie72a6b5e25845aa9b861dfab5aa7d315b5ce8b71