shill: cellular: remove unnecessary 'void' in function parameter lists
In C++, a function with an empty parameter list takes no argument, which
is different from C that requires using 'void' in the parameter list to
indicate that the function takes no argument. This CL removes
unnecessary 'void' in the parameter list for those functions that take
no argument.

BUG=None
TEST=Run unit tests.

Change-Id: I6ec8c28a4ba874e360de507b2dbbe8f762201a0a
Reviewed-on: https://chromium-review.googlesource.com/615098
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>