shill: wifi: remove unnecessary 'void' in function parameter lists
In C++, a function with an empty parameter list takes no argument, which
is different from C that requires using 'void' in the parameter list to
indicate that the function takes no argument. This CL removes
unnecessary 'void' in the parameter list for those functions that take
no argument.

BUG=None
TEST=Run unit tests.

Change-Id: I8e0868ec53301969bf47789deaae20e6b2d82fba
Reviewed-on: https://chromium-review.googlesource.com/615099
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>