shill: remove unnecessary 'void' in function parameter lists
In C++, a function with an empty parameter list takes no argument, which
is different from C that requires using 'void' in the parameter list to
indicate that the function takes no argument. This CL removes
unnecessary 'void' in the parameter list for those functions that take
no argument.

BUG=None
TEST=Run unit tests.

Change-Id: I2e53cfc7c68a386add1f5330e07b278d9146a85a
Reviewed-on: https://chromium-review.googlesource.com/616232
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>