shill: cellular: turn Cellular.SIMLockStatus.RetriesLeft into int32
Chrome converts shill properties into base::Value. Due to the lack of
uint32 support in base::Value, a shill property of type uint32 is held
as a double in base::Value, which often requires special handling on the
Chrome side to convert back and forth between double and uint32. That
incurs unnecessary complexity and makes code more error-prone.

This CL turns the 'RetriesLeft' field in the 'Cellular.SIMLockStatus'
device property from an uint32 value to an int32 value, which simplifies
the handling on the Chrome side.

BUG=chromium:753470
TEST=Run unit tests.
TEST=Manually perform SIM unlock from the Chrome settings UI and observe
the Cellular.SIMLockStatus property reported in chrome://network.

Change-Id: I8f368ea4e1a094018fc5bbca19e6fbc328646c28
Reviewed-on: https://chromium-review.googlesource.com/615729
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>