login: Write file when aborting browser.
This change writes a file to tmpfs when killing a browser process due to
a hang. This file can be detected (and deleted) by the crash-reporter.
The reporter can then annotate that crash as the result of a browser
hang.

This is simpler and more straightforward than parsing the log message
back out of /var/log/messages.

BUG=chromium:322185
TEST=Forced browser hang, confirmed file was written, also unit tests.

Change-Id: I8431252e76fc0b2d1e04f39b1f7f80608b283064
Reviewed-on: https://chromium-review.googlesource.com/617469
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>