shill: Make PortalURL a fixed string
Currently, PortalURL is a configurable Manager property so it is
cached in the default profile.  This means that if we want to change
the URL in the future, we cannot do it merely by changing the shill
code, because the cached value was set during the previous powerwash.

Nobody is currently making use of the ability to change the PortalURL
and save the setting on disk, however.  It was apparently left
over from flimflam.  All current shill users just use the default
hard-coded URL.  So, change shill so that the property is now
read-only, and merely returns the URL that is hard-coded into shill
rather than reading it from disk.

BUG=chromium:702273
TEST=unit tests
TEST=manually query Manager property and compare to on-disk profile
TEST=verify that captive portal check uses the hard-coded URL
CQ-DEPEND=CL:619590

Change-Id: I0d7c074243fef46cf01d4696c8489dbf72375ff2
Reviewed-on: https://chromium-review.googlesource.com/618223
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>