mtpd: add gyp file and move files to canonical locations
The gyp file is easier to understand than the makefile and will
make it easier to refactor. Also, move D-Bus configuration to
separate directories much like it is in other services that use
D-Bus and remove the old makefile since we're not using it
anymore.

CQ-DEPEND=CL:624910
BUG=chromium:757584
TEST=emerge-cyan mtpd; FEATURES=test emerge-cyan mtpd

Change-Id: I56121be11221aea5718f0a0daf012e29abc88a0c
Reviewed-on: https://chromium-review.googlesource.com/624859
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>