init: Re-order deletion in clobber-state
Start with clearing the TPM (the most effective means of destroying user
data), then progress to overwriting stateful and only after that get to
work on the non-active kernel and root partitions.

Fixes minor bug that VPD keys are not deleted in case the device loses
power at the wrong moment in time.

The previous deletion ordering prioritized preservation of state to
ensure that enterprise enrollment works seamlessly even in case of
aborted powerwash. However from a privacy perspective, the quick
deletion of data should have priority and it seems reasonable to place
the minor burden of having to run another powerwash before being able to
re-enroll on users that power off during powerwash. (The time window for
getting into an inconsistent state is short anyways.)

BUG=None
TEST=manual

Change-Id: I58637480c9d9eba1d05f7ed3673122e43d69b968
Reviewed-on: https://chromium-review.googlesource.com/629059
Commit-Ready: Thiemo Nagel <tnagel@chromium.org>
Tested-by: Thiemo Nagel <tnagel@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>