login: shutdown the android container more quickly
When the android container is booted during the login screen, it has no
user state. This allows us to kill it in a "quick and dirty" fashion
when the it is determined that the user has opted out of ARC. The
benefit of doing this is that it reduced the load on the system during a
performance critical time (tab restore).

BUG=b:62463418
TEST=Verified that new log message was seen. Observed the time to
shutdown the Android container halfed (0.88s to 0.43s). Ran
chromeos-login unit tests on caroline.

Change-Id: I5e0774f783954e59c532a964a243a66d279d14e6
Reviewed-on: https://chromium-review.googlesource.com/627638
Commit-Ready: Christopher Morin <cmtm@google.com>
Tested-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>